requirements.txt
azure-communication-callautomation
azure-identity
flask
azure-cognitiveservices-speech